added largefile support

+++ snet.h
@@ -35,8 +35,8 @@ int	snet_writef ___P(( SNET *, char *, ... ));
 char	*snet_getline ___P(( SNET *, struct timeval * ));
 char	*snet_getline_multi ___P(( SNET *, void (*)(char *),
 		struct timeval * ));
-int	snet_read ___P(( SNET *, char *, int, struct timeval * ));
-int	snet_write ___P(( SNET *, char *, int, struct timeval * ));
+ssize_t	snet_read ___P(( SNET *, char *, size_t, struct timeval * ));
+ssize_t	snet_write ___P(( SNET *, char *, size_t, struct timeval * ));
 #ifdef TLS
 int	snet_starttls ___P(( SNET *, SSL_CTX *, int ));
 #endif /* TLS */
